refactor: move app into `app/index.ts`
